Split off ResponseCookieTest
